requirements.txt
asn1crypto==1.3.0
autopep8==1.5
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
Django==2.1.5
django-braces==1.14.0
django-model-utils==4.0.0
django-multiselectfield==0.1.12
django-registration-redux==2.2
django-star-ratings==0.9.1
django-taggit==0.24.0
idna==2.8
lorem==0.1.1
lz4==3.0.2
olefile==0.46
Pillow==7.0.0
pycodestyle==2.5.0
pycparser==2.19
pyOpenSSL==19.1.0
PySocks==1.7.1
pytz==2019.3
requests==2.22.0
six==1.14.0
swapper==1.1.2.post1
urllib3==1.25.8
